backend/requirements.txt
# MedScribe AI - Backend Dependencies
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
greenlet==3.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.149.0

# HTTP
httpx==0.27.0
python-multipart==0.0.12

# Cache & Queue
redis==5.1.0
celery[redis]==5.4.0

# Storage (S3/MinIO)
boto3==1.35.0

# AI
anthropic==0.37.0
openai==1.50.0

# Search
meilisearch==0.31.0

# Encryption
cryptography==43.0.0

# Logging & Monitoring
structlog==24.4.0
sentry-sdk[fastapi]==2.14.0
prometheus-client==0.21.0

# Rate Limiting
slowapi==0.1.9

# PDF
reportlab==4.2.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0

# Dev tools
ruff==0.6.0
mypy==1.11.0
